backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
pdfplumber==0.11.0
pypdf==4.2.0
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
PyYAML==6.0.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
openai==1.35.3
